#18Jun2025
Enhance form handling and validation logic

Updated methods in FormBuilderController and FormController to improve null checks, error handling, and the assignment of fields and approvers. Enhanced form validation in DynamicForm.jsx and added a required checkbox in FormField component. Introduced a new `Required` property in GridColumnDTO and FormField for better data representation.